Revert "Merge pull request #18148 from Budibase/revert-remove-old-llm-code"
This reverts commit c56edbccd2cd303b8c8f624082fc6bbf73327eda, reversing
changes made to 801b285d7d227a1ce43c0fb9a207c7fa0e02caf4.
