+++ packages/backend-core/package.json
@@ -41,7 +41,6 @@
     "@budibase/types": "*",
     "@govtechsg/passport-openidconnect": "1.0.3",
     "aws-cloudfront-sign": "3.0.2",
-    "aws-sdk": "2.1692.0",
     "bcrypt": "6.0.0",
     "bull": "4.10.1",
     "correlation-id": "4.0.0",
