+++ packages/pro/src/ai/index.ts
@@ -1,8 +1,6 @@
 export * from "./utils"
 export * from "./prompts"
-export { parseResponseFormat } from "./models/openai"
-export { getLLM, getLLMConfig, LLMRequest } from "./llm"
-export { LLM } from "./models/base"
+export { getLLMConfig, LLMRequest, parseResponseFormat } from "./llm"
 export * from "./prompts"
 export * from "./reasoningFallback"
 export * from "./generators"
