--- packages/pro/src/ai/models/anthropic.ts
@@ -1,99 +1,0 @@
-import AnthropicClient from "@anthropic-ai/sdk"
-import { LLMConfigOptions } from "@budibase/types"
-import { LLMFullResponse } from "../../types/ai"
-import { LLMRequest } from "../llm"
-import { LLM } from "./base"
-
-export type AnthropicModel =
-  | "claude-3-5-sonnet-20240620"
-  | "claude-3-sonnet-20240229"
-  | "claude-3-opus-20240229"
-  | "claude-3-haiku-20240307"
-
-function calculateBudibaseAICredits(usage?: AnthropicClient.Usage): number {
-  if (!usage) {
-    return 0
-  }
-
-  // Output tokens are 3 times more expensive in OpenAI
-  // We want to have a single figure in BB for simplicity
-  const inputTokens = usage.input_tokens
-  const outputTokens = usage.output_tokens
-  return outputTokens * 3 + inputTokens
-}
-
-export class Anthropic extends LLM {
-  override supportsFiles = false
-
-  private client: AnthropicClient
-
-  constructor(opts: LLMConfigOptions) {
-    super(opts)
-    this.client = new AnthropicClient({ apiKey: opts.apiKey })
-  }
-
-  firstTextBlock(message: AnthropicClient.Messages.Message) {
-    for (const content of message.content) {
-      if (content.type !== "text") {
-        continue
-      }
-      return content.text
-    }
-    return undefined
-  }
-  protected async chatCompletion(
-    request: LLMRequest
-  ): Promise<LLMFullResponse> {
-    try {
-      const completion = await this.client.messages.create({
-        model: this.model,
-        messages: request.messages.map(({ content }) => {
-          if (content == null) {
-            return { role: "user", content: "" }
-          }
-          if (typeof content === "string") {
-            return {
-              // Anthropic's API doesn't seem to have any notion of system messages,
-              // just "user" and "assistant" where "assistant" represents the AI.
-              role: "user",
-              content,
-            }
-          }
-
-          let message = ""
-          for (const part of content) {
-            if (part.type === "text") {
-              message += part.text
-            } else if (
-              ["image_url", "input_audio", "file"].includes(part.type)
-            ) {
-              // TODO: maybe support this one day
-            }
-          }
-
-          return {
-            role: "user",
-            content: message,
-          }
-        }),
-        max_tokens: this.maxTokens,
-      })
-
-      const message = this.firstTextBlock(completion) || ""
-      return {
-        messages: [
-          ...request.messages,
-          { role: "assistant", content: message },
-        ],
-        tokensUsed: calculateBudibaseAICredits(completion.usage),
-      }
-    } catch (err) {
-      if (err instanceof AnthropicClient.APIError) {
-        console.error(
-          `Anthropic Prompt failed with error ${err.name} and status: ${err.status}`
-        )
-      }
-      throw err
-    }
-  }
-}
