--- packages/pro/src/ai/models/index.ts
@@ -1,2 +1,0 @@
-export * from "./openai"
-export * from "./anthropic"
