+++ packages/server/package.json
@@ -44,7 +44,6 @@
     "@ai-sdk/azure": "^3.0.30",
     "@ai-sdk/openai": "^3.0.26",
     "@ai-sdk/provider": "^3.0.8",
-    "@anthropic-ai/sdk": "^0.27.3",
     "@apidevtools/swagger-parser": "12.1.0",
     "@aws-sdk/client-dynamodb": "3.709.0",
     "@aws-sdk/client-s3": "3.709.0",
@@ -53,7 +52,6 @@
     "@azure/msal-node": "^2.5.1",
     "@budibase/backend-core": "*",
     "@budibase/client": "*",
-    "@budibase/frontend-core": "*",
     "@budibase/nano": "10.1.5",
     "@budibase/pro": "*",
     "@budibase/shared-core": "*",
@@ -69,7 +67,6 @@
     "@google-cloud/firestore": "7.8.0",
     "@koa/cors": "5.0.0",
     "@koa/router": "15.3.0",
-    "@octokit/rest": "^20.0.0",
     "@socket.io/redis-adapter": "^8.2.1",
     "@types/xml2js": "^0.4.14",
     "ai": "^6.0.3",
@@ -78,7 +75,6 @@
     "archiver": "7.0.1",
     "bcrypt": "6.0.0",
     "bson": "^6.9.0",
-    "buffer": "6.0.3",
     "bull": "4.10.1",
     "chat": "^4.14.0",
     "chokidar": "3.5.3",
@@ -90,7 +86,6 @@
     "dd-trace": "5.63.0",
     "dotenv": "8.2.0",
     "extract-zip": "^2.0.1",
-    "form-data": "4.0.4",
     "global-agent": "3.0.0",
     "google-auth-library": "^10.5.0",
     "google-spreadsheet": "npm:@budibase/google-spreadsheet@4.1.5",
@@ -125,12 +120,10 @@
     "pdf-parse": "^2.4.5",
     "pg": "8.10.0",
     "pouchdb": "9.0.0",
-    "pouchdb-all-dbs": "1.1.1",
     "pouchdb-find": "9.0.0",
     "redis": "4",
     "semver": "^7.5.4",
     "serialize-error": "^7.0.1",
-    "server-destroy": "1.0.1",
     "snowflake-sdk": "^1.15.0",
     "socket.io": "4.8.1",
     "svelte": "^5.53.5",
@@ -152,7 +145,6 @@
     "@types/global-agent": "2.1.1",
     "@types/html-to-text": "^9.0.4",
     "@types/js-yaml": "^4.0.9",
-    "@types/json-schema": "^7.0.15",
     "@types/koa": "2.13.4",
     "@types/koa-send": "^4.1.6",
     "@types/koa__cors": "5.0.0",
@@ -164,7 +156,6 @@
     "@types/pg": "8.6.6",
     "@types/pouchdb": "6.4.2",
     "@types/semver": "7.3.7",
-    "@types/server-destroy": "1.0.1",
     "@types/supertest": "2.0.14",
     "@types/swagger-jsdoc": "^6.0.4",
     "@types/tar": "6.1.5",
