+++ packages/string-templates/package.json
@@ -24,8 +24,7 @@
     "@budibase/handlebars-helpers": "^0.13.2",
     "@budibase/vm-browserify": "^1.1.4",
     "dayjs": "^1.10.8",
-    "handlebars": "^4.7.8",
-    "lodash.clonedeep": "^4.5.0"
+    "handlebars": "^4.7.8"
   },
   "devDependencies": {
     "@rollup/plugin-commonjs": "^17.1.0",
