+++ packages/worker/package.json
@@ -37,7 +37,6 @@
     "@koa/router": "15.3.0",
     "@govtechsg/passport-openidconnect": "1.0.3",
     "@types/global-agent": "2.1.1",
-    "aws-sdk": "2.1692.0",
     "bcrypt": "6.0.0",
     "bull": "4.10.1",
     "dd-trace": "5.63.0",
@@ -48,28 +47,19 @@
     "ical-generator": "4.1.0",
     "joi": "17.6.0",
     "jsonwebtoken": "9.0.2",
-    "knex": "2.4.2",
     "koa": "2.15.4",
     "koa-body": "4.2.0",
     "koa-compress": "4.0.1",
-    "koa-passport": "4.1.4",
     "koa-redis": "^4.0.1",
-    "koa-send": "5.0.1",
     "koa-session": "5.13.1",
-    "koa-static": "5.0.0",
     "koa-useragent": "^4.1.0",
     "lodash": "4.17.23",
     "marked": "^15.0.11",
     "node-fetch": "2.6.7",
     "nodemailer": "7.0.11",
-    "passport-google-oauth": "2.0.0",
-    "passport-local": "1.0.0",
     "pouchdb": "9.0.0",
-    "pouchdb-all-dbs": "1.1.1",
     "scim2-parse-filter": "^0.2.8",
     "scim-patch": "^0.8.1",
-    "server-destroy": "1.0.1",
-    "undici": "^7.16.0",
     "uuid": "^8.3.2",
     "yaml": "^2.8.2"
   },
@@ -79,7 +69,6 @@
     "@types/maildev": "^0.0.7",
     "@types/node-fetch": "2.6.4",
     "@types/nodemailer": "^6.4.17",
-    "@types/server-destroy": "1.0.1",
     "@types/supertest": "2.0.14",
     "@types/uuid": "8.3.4",
     "cheerio": "^1.0.0",
